+++ models/posts.js
@@ -32,9 +32,9 @@ const PostsSchema           = new mongoose.Schema({
     category                : { type: String, lowercase: true, default: 'unclassified' },
     tags                    : { type: String, lowercase: true },
     visibility: {
-        hidden              : { type: Boolean, required: true, default: false },    // todo: add anti-robot HTML tag
-        pinned              : { type: Boolean, required: true, default: false },
-        protected           : { type: Boolean, required: true, default: false },    // todo: add pw
+        hidden              : { type: Boolean, default: false },    // todo: add anti-robot HTML tag
+        pinned              : { type: Boolean, default: false },
+        protected           : { type: Boolean, default: false },    // todo: add pw
     },
     status                  : { type: String, default: 'published',
                                 enum: ['drafted','published', 'recycled'] },
@@ -53,18 +53,6 @@ const PostsSchema           = new mongoose.Schema({
 // ==============================
 //  STATIC METHODS
 // ==============================
-// create and associate (model)
-PostsSchema.static('postsCreateThenAssociate', function (raw, user, next) {
-    return Promise.resolve().then(() => _U_.schema.updateAndBind(raw, user, next, 'posts', '$push', this));
-});
-
-
-// delete and dissociate (model)  // note: not workable for admin in deleting media owned by multiple users
-PostsSchema.static('postsDeleteThenDissociate', function (docsID, user, next) {
-    return Promise.resolve().then(() => _U_.schema.updateAndBind(docsID, user, next, 'posts', '$pullAll', this));
-});
-
-
 // (pre-hook) version counter
 PostsSchema.pre('findOneAndUpdate', function () {
     this.findOneAndUpdate({}, { $inc: { _revised: 1 }});